standalone-utils/dashboard/requirements.txt
# Core dependencies for XEMM Dashboard
hyperliquid-python-sdk
websockets
aiohttp
python-dotenv>=1.0.0
rich>=13.0.0
solders
base58
eth-account
requests>=2.25.0
typing-extensions>=4.0.0
